+++ examples/src/bin/embassy_multicore_interrupt.rs
@@ -91,9 +91,9 @@ fn main() -> ! {
     let io = Io::new(peripherals.GPIO, peripherals.IO_MUX);
 
     let timg0 = TimerGroup::new(peripherals.TIMG0, &clocks);
-    let timer0 = OneShotTimer::new(timg0.timer0.into());
-    let timer1 = OneShotTimer::new(timg0.timer1.into());
-    let timers = [timer0, timer1];
+    let timer0: ErasedTimer = timg0.timer0.into();
+    let timer1: ErasedTimer = timg0.timer1.into();
+    let timers = [OneShotTimer::new(timer0), OneShotTimer::new(timer1)];
     let timers = mk_static!([OneShotTimer<ErasedTimer>; 2], timers);
     esp_hal_embassy::init(&clocks, timers);